+++ partial_this/tests/crate_name.rs
@@ -14,9 +14,9 @@ mod default_crate_name {
     #[test]
     fn builds_partially() {
         let p = Foo::partial(Box::new_uninit());
-        let mut a = p.foo(1);
+        let mut a = p.with_foo(1);
         a.set_foo(123);
-        let r = a.bar(456.0).done();
+        let r = a.with_bar(456.0).done();
         assert_eq!(r.foo, 123);
         assert_eq!(r.bar, 456.0);
     }
@@ -34,9 +34,9 @@ mod ident_crate_name {
     #[test]
     fn builds_partially() {
         let p = Foo::partial(Box::new_uninit());
-        let mut a = p.foo(1);
+        let mut a = p.with_foo(1);
         a.set_foo(123);
-        let r = a.bar(456.0).done();
+        let r = a.with_bar(456.0).done();
         assert_eq!(r.foo, 123);
         assert_eq!(r.bar, 456.0);
     }
@@ -54,9 +54,9 @@ mod string_crate_name {
     #[test]
     fn builds_partially() {
         let p = Foo::partial(Box::new_uninit());
-        let mut a = p.foo(1);
+        let mut a = p.with_foo(1);
         a.set_foo(123);
-        let r = a.bar(456.0).done();
+        let r = a.with_bar(456.0).done();
         assert_eq!(r.foo, 123);
         assert_eq!(r.bar, 456.0);
     }