+++ partial_this/tests/scopes.rs
@@ -15,7 +15,7 @@ mod module_name_conflict {
     #[test]
     fn builds_with_conflicting_module_name() {
         let p = Foo::partial(Box::new_uninit());
-        let foo = p.a(1).done();
+        let foo = p.with_a(1).done();
         assert_eq!(foo.a, 1);
     }
 }
@@ -35,7 +35,7 @@ mod module_qualified_field_type {
     #[test]
     fn builds_with_module_qualified_field_type() {
         let p = Holder::partial(Box::new_uninit());
-        let h = p.v(inner::Value(7)).done();
+        let h = p.with_v(inner::Value(7)).done();
         assert_eq!(h.v.0, 7);
     }
 }
@@ -53,7 +53,7 @@ mod custom_local_field_type {
     #[test]
     fn builds_with_local_field_type() {
         let p = Blob::partial(Box::new_uninit());
-        let b = p.data(Local(9)).done();
+        let b = p.with_data(Local(9)).done();
         assert_eq!(b.data.0, 9);
     }
 }
@@ -70,7 +70,7 @@ mod all_pub {
 
     #[test]
     fn builds() {
-        let foo = Foo::partial(Box::new_uninit()).a(1).b(2.0).done();
+        let foo = Foo::partial(Box::new_uninit()).with_a(1).with_b(2.0).done();
         assert_eq!((foo.a, foo.b), (1, 2.0));
     }
 }
@@ -87,7 +87,7 @@ mod mixed_visibility {
 
     #[test]
     fn builds() {
-        let foo = Foo::partial(Box::new_uninit()).a(1).b(2.0).done();
+        let foo = Foo::partial(Box::new_uninit()).with_a(1).with_b(2.0).done();
         assert_eq!((foo.a, foo.b), (1, 2.0));
     }
 }
@@ -103,7 +103,7 @@ mod pub_use_false_config {
 
     #[test]
     fn builds() {
-        let foo = Foo::partial(Box::new_uninit()).a(1).done();
+        let foo = Foo::partial(Box::new_uninit()).with_a(1).done();
         assert_eq!(foo.a, 1);
     }
 }
@@ -119,7 +119,7 @@ mod pub_use_false_private_struct {
 
     #[test]
     fn builds() {
-        let foo = Foo::partial(Box::new_uninit()).a(1).done();
+        let foo = Foo::partial(Box::new_uninit()).with_a(1).done();
         assert_eq!(foo.a, 1);
     }
 }
@@ -135,7 +135,7 @@ mod pub_use_false_pub_crate_struct {
 
     #[test]
     fn builds() {
-        let foo = Foo::partial(Box::new_uninit()).a(1).done();
+        let foo = Foo::partial(Box::new_uninit()).with_a(1).done();
         assert_eq!(foo.a, 1);
     }
 }
@@ -151,7 +151,7 @@ mod pub_crate_struct {
 
     #[test]
     fn builds() {
-        let foo = Foo::partial(Box::new_uninit()).a(1).done();
+        let foo = Foo::partial(Box::new_uninit()).with_a(1).done();
         assert_eq!(foo.a, 1);
     }
 }
@@ -168,7 +168,7 @@ mod private_struct {
 
     #[test]
     fn builds() {
-        let foo = Foo::partial(Box::new_uninit()).a(1).done();
+        let foo = Foo::partial(Box::new_uninit()).with_a(1).done();
         assert_eq!(foo.a, 1);
     }
 }
@@ -187,7 +187,10 @@ mod private_field_type {
 
     #[test]
     fn builds() {
-        let foo = Foo::partial(Box::new_uninit()).a(Hidden(3)).b(1).done();
+        let foo = Foo::partial(Box::new_uninit())
+            .with_a(Hidden(3))
+            .with_b(1)
+            .done();
         let Foo { a, b } = *foo;
         assert_eq!(a.0, 3);
         assert_eq!(b, 1);