+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2026-09-03
+
+### Changed
+
+- Improve the field-setting API: rename the field setter from `field(value)` to
+  `with_field(value)`, and add `emplace_field(init)` and `uninit_field()` for
+  in-place initialization of a field.
+- Rename the accessor `get_field_mut()` to `get_mut_field()`, and make
+  `set_field()` return `&mut Self` for chaining.
+
 ## [0.3.1] - 2026-09-03
 
 ### Added